Update for factory
