+++ src/Maui/Features/Creatures/Import/Upload.razor.cs
@@ -1,4 +1,4 @@
-﻿using DMPowerTools.Core.Features.Creatures;
+﻿using DMPowerTools.Core.Features.Creatures.Import;
 
 namespace DMPowerTools.Maui.Features.Creatures.Import;
 
@@ -7,9 +7,9 @@ public partial class Upload
     [Inject] private IMediator Mediator { get; set; } = null!;
     [CascadingParameter] public ImportState ImportState { get; set; }
 
-    private async Task UploadFilesAsync(IReadOnlyList<IBrowserFile> files, ImportSource importSource)
+    private async Task UploadFilesAsync(IReadOnlyList<IBrowserFile> files, UploadSource uploadSource)
     {
-        var response = await Mediator.Send(new ProcessMonsterFileCommand(files));
+        var response = await Mediator.Send(new UploadMonsterFileCommand(files, uploadSource));
 
         ImportState.InReviewCreatures.AddRange(response.Creatures);
